refactor is_onepass to be part of Hir construction.
