+++ src/onepass.rs
@@ -27,7 +27,6 @@ use prog::{Program, Inst, EmptyLook};
 use literal::LiteralSearcher;
 use re_trait::Slot;
 use input::{ByteInput, Input};
-use analysis;
 use compile::Compiler;
 use syntax::hir::Hir;
 use re_builder::RegexOptions;
@@ -530,7 +529,7 @@ impl OnePassCompiler {
             return Err(OnePassError::RegexSetUnsupported);
         }
 
-        if ! analysis::is_onepass(&es[0]) {
+        if ! es[0].is_onepass() {
             return Err(OnePassError::HasNondeterminism);
         }
